refactor: decouple test identity row type
