+++ bitloops/src/capability_packs/test_harness/event_handlers.rs
@@ -5,8 +5,6 @@ mod persistence;
 #[cfg(test)]
 mod tests;
 
-pub(crate) use self::persistence::ExistingTestArtefactIdentityRow;
-
 use crate::host::capability_host::{
     CurrentStateConsumer, CurrentStateConsumerContext, CurrentStateConsumerFuture,
     CurrentStateConsumerRequest, CurrentStateConsumerResult, ReconcileMode,
